Add Dragula library and create basic page layout
This was a bit much on the commit, because I let too much time go
between working on it, and I forgot I was trying to keep track of all
the commands and all little changes.

This commit adds the Dragula library to handle the drag and drop. I
spent way too much time looking at all the drag-n-drop libraries and
playing with the native HTML5 capabilities. Dragula seems the most
active. There is also (interact.js)[http://interactjs.io/], but that
wasn't working for various reasons. Also the ability to clone on drag
was too involved. Dragula seems more simple to implement with less
written code.

This commit also creates the basic layout of the application. Two
columns with days of the week on the left and a list of recipes on the
right. The recipes are pulled from a static JSON file that was generated
from some recipe website that I can't remember anymore. There are
APIs that provide such data, and I will use them later.

This commit, then, creates the basic layout, and gets the basic
functionality of dragging a recipe onto the day. The recipe is cloned,
so that the original stays in the recipe list and can be used for
multiple days.

STEPS
- Install the Dragula library to handle the drag and drop functionality
  - `npm install --save dragula`
  - this updates the package.json and package-lock.json files
- Make symbolic link for the dragula scripts
  - Fist need to be in the correct website directory:
    - `cd public/stylesheets/`
    - Then create the symbolic link (short cut)
      - `ln -s ../../node_modules/dragula/dist/dragula.css dragula.css`
  - Next for the javascript
    - `cd ../javascripts/`
    - `ln -s ../../node_modules/dragula/dist/dragula.min.js
      dragula.min.js`
- Create a javascript file for handling the drag and drop, and for
  creating the list of recipes (temporarily handled this way).
  - create the file in 'public/javascripts/mealer.js'
- Create a JSON file to hold a list of recipes to use for now
 - create the file in 'public/recipies.json'
- Edit the 'views/index.pug' file to have two columns, days of the week
  on the left, each day a different color, and the list of menu items on
  the right.
- Edit the 'views/layout.pug' file to include all of the stylesheets
- Create a new 'views/includes' directory (if in the root directory of
  the project)
  - `mkdir views/includes`
- Create a new file in 'views/includes/foot.pug' to hold the footer information, which will be
  included in the layout (or any page that needs it).
- Update the stylesheet 'public/stylesheets/style.css' make the image in the recipe list a max of 100
  pixels (otherwise they are absolutely huge), and to make the div for
  the day large enough for the recipe to be dragged onto it.