Redo the layout for simplicity and functionality.
This commit restructures the layout following the example on the
materializeCSS site for the Example Side Navigation Layout; one row for
the whole page, the first column spans all 12 columns, the next two
split the columns 4/8.

This commit takes out the styles to the body tag that are all unneeded.
It adds a min height to the day class, which is an empty div for
receiving the recipes when dragged, adds some padding and styles to the
search at the top to make it more visibly pleasing, and constrains the
recipe image to 100 pixels high.