Allows the window to scroll while dragging.
This commit includes the dom_autoscroll library which adds the
functionality for the window to scroll while the recipe is being
dragged. Before, several days were out of view, and there was no way to
drag a recipe from the top of the menu list to the day at the bottom of
the week list. This library allows for this behavior.

It also cleans up the structure in the 'views/index.pug' file by putting
the ID first in the divs, with the column size, next, then colors, then
other classes. It also moves the color to the card div, instead of the
card-content div, and removes the container class on the row div that
holds the days. This makes the day card much larger. It also tweaks the
sizing of the week and meals columns to a 5/7 ratio so that all of the
text on a recipe card is contained when on small screens.

STEPS:
- Install the dom-autoscroller library
  - `npm install --save dom-autoscroller`
- Make a symbolic link for the dom-autoscroller script.
  - Change directory into the 'public/javascripts' directory.
    - `ln -s
      ../../node_modules/dom-autoscroller/dist/dom-autoscroller.min.js
      dom-autoscroller.min.js`
- Import the script in the 'views/includes/foot.pug' file.
- Modify the 'public/javascripts/mealer.js' file to use the
  dom-autoscroller functionality.
  - Another change removes the unwanted 'truncate' class from the recipe
    title span.