Add delete button to cloned meal card
This commit first reformats how the drake is written to make it more
legible. It then adds the .on method with the drop identifier, which
then calls the 'addDeleteMealButton' function. This function adds an
icon to the end of the card div, which has an onclick event that will
delete the card.