Move creation of days to mealer.js and days.json files
This moves the data about the days cards to a days.json file which for
now just holds the day of the week and the color of the card. Then
mealer.js dynamically creates the day cards using the days.json file.
The index.pug file is updated to not have the lines of code that
formerly created the day cards, and adds a pug-only comment that the day
and meal cards are generated in the mealer.js file.