Fixes the display of the day for each day card.
- This uses the default JavaScript ability to call out parts of a Date
object and format them how you wish. I want the day of the week in long
name format, the full month name, and the day of the month in number.
Passing toLocaleString to the date object (dday) can get me each of
those.

- deletes the unnecessary array of names of the days of the week.

- this also takes out hard-coded numbers from several spots and moves
them up to variables at the top of the getDaysList function. Much better
coding practice. Eventually, all of these will be set in a user's
account settings.